Rename CreateSimpleBrushTool to DrawBrushTool
