+++ app/resources/documentation/manual/images/DrawBrush.gif
