+++ common/src/View/DrawBrushToolController2D.cpp
@@ -17,15 +17,10 @@
  along with TrenchBroom. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "CreateSimpleBrushToolController3D.h"
+#include "DrawBrushToolController2D.h"
 
-#include "FloatType.h"
-#include "Model/BrushNode.h"
-#include "Model/Hit.h"
-#include "Model/HitFilter.h"
-#include "Model/PickResult.h"
 #include "Renderer/Camera.h"
-#include "View/CreateSimpleBrushTool.h"
+#include "View/DrawBrushTool.h"
 #include "View/Grid.h"
 #include "View/HandleDragTracker.h"
 #include "View/InputState.h"
@@ -33,43 +28,44 @@
 
 #include <kdl/memory_utils.h>
 
-#include <vecmath/bbox.h>
-#include <vecmath/line.h>
-#include <vecmath/plane.h>
-#include <vecmath/vec.h>
+#include <vecmath/intersection.h>
+#include <vecmath/scalar.h>
 
 namespace TrenchBroom::View
 {
 
-CreateSimpleBrushToolController3D::CreateSimpleBrushToolController3D(
-  CreateSimpleBrushTool& tool, std::weak_ptr<MapDocument> document)
+DrawBrushToolController2D::DrawBrushToolController2D(
+  DrawBrushTool& tool, std::weak_ptr<MapDocument> document)
   : m_tool{tool}
   , m_document{std::move(document)}
 {
 }
 
-Tool& CreateSimpleBrushToolController3D::tool()
+Tool& DrawBrushToolController2D::tool()
 {
   return m_tool;
 }
 
-const Tool& CreateSimpleBrushToolController3D::tool() const
+const Tool& DrawBrushToolController2D::tool() const
 {
   return m_tool;
 }
 
 namespace
 {
-class CreateSimpleBrushDragDelegate : public HandleDragTrackerDelegate
+class DrawBrushDragDelegate : public HandleDragTrackerDelegate
 {
 private:
-  CreateSimpleBrushTool& m_tool;
+  DrawBrushTool& m_tool;
   vm::bbox3 m_worldBounds;
+  vm::bbox3 m_referenceBounds;
 
 public:
-  CreateSimpleBrushDragDelegate(CreateSimpleBrushTool& tool, const vm::bbox3& worldBounds)
+  DrawBrushDragDelegate(
+    DrawBrushTool& tool, const vm::bbox3& worldBounds, const vm::bbox3& referenceBounds)
     : m_tool{tool}
     , m_worldBounds{worldBounds}
+    , m_referenceBounds{referenceBounds}
   {
   }
 
@@ -83,31 +79,13 @@ public:
     m_tool.update(currentBounds);
     m_tool.refreshViews();
 
+    const auto& camera = inputState.camera();
+    const auto plane = vm::plane3{
+      initialHandlePosition,
+      vm::vec3{vm::get_abs_max_component_axis(camera.direction())}};
+
     return makeHandlePositionProposer(
-      makePlaneHandlePicker(vm::horizontal_plane(initialHandlePosition), handleOffset),
-      makeIdentityHandleSnapper());
-  }
-
-  std::optional<UpdateDragConfig> modifierKeyChange(
-    const InputState& inputState, const DragState& dragState) override
-  {
-    if (inputState.modifierKeys() == ModifierKeys::MKAlt)
-    {
-      return UpdateDragConfig{
-        makeHandlePositionProposer(
-          makeLineHandlePicker(
-            vm::line3{dragState.currentHandlePosition, vm::vec3::pos_z()},
-            dragState.handleOffset),
-          makeIdentityHandleSnapper()),
-        ResetInitialHandlePosition::Keep};
-    }
-
-    return UpdateDragConfig{
-      makeHandlePositionProposer(
-        makePlaneHandlePicker(
-          vm::horizontal_plane(dragState.currentHandlePosition), dragState.handleOffset),
-        makeIdentityHandleSnapper()),
-      ResetInitialHandlePosition::Keep};
+      makePlaneHandlePicker(plane, handleOffset), makeIdentityHandleSnapper());
   }
 
   DragStatus drag(
@@ -166,56 +144,36 @@ private:
     const vm::vec3& initialHandlePosition,
     const vm::vec3& currentHandlePosition) const
   {
-    const auto bounds = vm::bbox3{
-      vm::min(initialHandlePosition, currentHandlePosition),
-      vm::max(initialHandlePosition, currentHandlePosition)};
+    const auto bounds = vm::merge(
+      vm::bbox3{initialHandlePosition, initialHandlePosition}, currentHandlePosition);
     return vm::intersect(snapBounds(inputState, bounds), m_worldBounds);
   }
 
-  vm::bbox3 snapBounds(const InputState& inputState, vm::bbox3 bounds) const
+  vm::bbox3 snapBounds(const InputState& inputState, const vm::bbox3& bounds) const
   {
-
-    // prevent flickering due to very small rounding errors
-    bounds.min = vm::correct(bounds.min);
-    bounds.max = vm::correct(bounds.max);
-
     const auto& grid = m_tool.grid();
-    bounds.min = grid.snapDown(bounds.min);
-    bounds.max = grid.snapUp(bounds.max);
+    auto min = grid.snapDown(bounds.min);
+    auto max = grid.snapUp(bounds.max);
 
     const auto& camera = inputState.camera();
-    const auto cameraPosition = vm::vec3{camera.position()};
+    const auto& refBounds = m_referenceBounds;
+    const auto factors =
+      vm::vec3{vm::abs(vm::get_abs_max_component_axis(camera.direction()))};
+    min = vm::mix(min, refBounds.min, factors);
+    max = vm::mix(max, refBounds.max, factors);
 
-    for (size_t i = 0; i < 3; i++)
-    {
-      if (bounds.max[i] <= bounds.min[i])
-      {
-        if (bounds.min[i] < cameraPosition[i])
-        {
-          bounds.max[i] = bounds.min[i] + grid.actualSize();
-        }
-        else
-        {
-          bounds.min[i] = bounds.max[i] - grid.actualSize();
-        }
-      }
-    }
-
-    return bounds;
+    return vm::bbox3{min, max};
   }
 };
 } // namespace
 
-std::unique_ptr<DragTracker> CreateSimpleBrushToolController3D::acceptMouseDrag(
+std::unique_ptr<DragTracker> DrawBrushToolController2D::acceptMouseDrag(
   const InputState& inputState)
 {
-  using namespace Model::HitFilters;
-
   if (!inputState.mouseButtonsPressed(MouseButtons::MBLeft))
   {
     return nullptr;
   }
-
   if (!inputState.modifierKeysPressed(ModifierKeys::MKNone))
   {
     return nullptr;
@@ -227,18 +185,27 @@ std::unique_ptr<DragTracker> CreateSimpleBrushToolController3D::acceptMouseDrag(
     return nullptr;
   }
 
-  const auto& hit = inputState.pickResult().first(type(Model::BrushNode::BrushHitType));
-  const auto initialHandlePosition =
-    hit.isMatch() ? hit.hitPoint() : inputState.defaultPointUnderMouse();
+  const auto& bounds = document->referenceBounds();
+  const auto& camera = inputState.camera();
+  const auto plane =
+    vm::plane3{bounds.min, vm::vec3{vm::get_abs_max_component_axis(camera.direction())}};
 
+  const auto distance = vm::intersect_ray_plane(inputState.pickRay(), plane);
+  if (vm::is_nan(distance))
+  {
+    return nullptr;
+  }
+
+  const auto initialHandlePosition =
+    vm::point_at_distance(inputState.pickRay(), distance);
   return createHandleDragTracker(
-    CreateSimpleBrushDragDelegate{m_tool, document->worldBounds()},
+    DrawBrushDragDelegate{m_tool, document->worldBounds(), document->referenceBounds()},
     inputState,
     initialHandlePosition,
     initialHandlePosition);
 }
 
-bool CreateSimpleBrushToolController3D::cancel()
+bool DrawBrushToolController2D::cancel()
 {
   return false;
 }
